# openapi.yaml
openapi: 3.0.0
info:
  title: Order Service
  description: Manages orders for the K11 platform. Consumes user-service to attach user profiles.
  version: 1.0.0
paths:
  /api/v1/orders:
    get:
      summary: List all orders
      operationId: list_orders
      tags: [orders]
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
    post:
      summary: Create a new order
      operationId: create_order
      tags: [orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrder'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '422':
          description: User not found
  /api/v1/orders/{id}:
    get:
      summary: Get an order by ID
      operationId: get_order
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
    patch:
      summary: Update order status
      operationId: update_order_status
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          required: true
          schema:
            type: string
            enum: [pending, confirmed, shipped, delivered, cancelled]
      responses:
        '200':
          description: Order updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
components:
  schemas:
    Order:
      type: object
      required: [id, user_id, items, total, status]
      properties:
        id:
          type: string
        user_id:
          type: string
          description: ID from user-service GET /api/v2/users/{id}
        items:
          type: array
          items:
            type: string
        total:
          type: number
          format: float
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
    CreateOrder:
      type: object
      required: [user_id, items, total]
      properties:
        user_id:
          type: string
        items:
          type: array
          items:
            type: string
        total:
          type: number
          format: float